Fill speaker photo URL and past experience from user profile
In case the user profile is not complete, we use the default value
'unknown' rather than silently failing.

A better solution would be to first ask the user to update its profile
(and make the photo URL and past experience fields mandatory in that
case) and then use the values she has just filled.